set X-Forwarded-For for forum.garudalinux.org
We parse the cloudflare IPs on our end and then pass the real IPs to the discourse container.
On discourse's end, there is a custom template that simply sets X-Forwarded-For as the real_ip_header, taking either the real ip from cloudflare or actual real IP as the real IP address

The reason the old method didn't work: 192.168.0.20 was not a trusted IP address. And manually adding that IP address to the trusted IP addresses would allow anyone that sets CF-Connecting-IP to claim to be any IP address even if they were not cloudflare.