Nested nspawn via nixos-containers at last
Why does this work? I have no clue. Why does remounting only work half the time? I have no clue. Why does this not work when done in a systemd service inside of the container? I have no clue. I have spent hours on this. If you read this and find this to be useful for you, you owe me.