feat(garuda-mail,garuda-build): set default SSH port to 1022
After noticing the insane amount of login spams, changing the SSH port to prevent the automated part of login attempts is a good way to block a portion of them.

Immortalis default port is already on 222, while temeraire (the Chaotic main node) needs to be on port 22 because of how Chaotic-AUR's toolbox works.